[MIN-475] move integration tests from liquidity_pools to controller
